\name{demo_cn}
\alias{demo_cn}
\docType{data}
\title{
Demo copy number data
}
\description{
A data frame containing partial Battenberg copy number data for demonstration.
}
\usage{data("demo_cn")}
\format{
  A data frame with 1 observations on the following 5 variables.
  \describe{
    \item{\code{chr}}{a numeric vector}
    \item{\code{startpos}}{a numeric vector}
    \item{\code{endpos}}{a numeric vector}
    \item{\code{nMaj1_A}}{a numeric vector}
    \item{\code{nMin1_A}}{a numeric vector}
  }
}
\source{
data(AmplificationTimeRdata)
}
\examples{
demo_cn
}
\keyword{datasets}
