\name{demo_mult}
\alias{demo_mult}
\docType{data}
\title{
Demo multiplicity data
}
\description{
A data frame containing partial dpclust3p multiplicity data for demonstration.
}
\usage{data("demo_mult")}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{chr}}{a character vector}
    \item{\code{end}}{a numeric vector}
    \item{\code{no.chrs.bearing.mut}}{a numeric vector}
  }
}
\examples{
demo_mult
}
\keyword{datasets}
