\name{demo_muts}
\alias{demo_muts}
\docType{data}
\title{
Demo mutation data
}
\description{
A data frame containing mutation data for demonstration. 
}
\usage{data("demo_muts")}
\format{
  A data frame with 20 observations on the following 5 variables.
  \describe{
    \item{\code{chr}}{a character vector}
    \item{\code{start}}{a numeric vector}
    \item{\code{end}}{a numeric vector}
    \item{\code{ref}}{a character vector}
    \item{\code{alt}}{a character vector}
  }
}

\examples{
demo_muts
}
\keyword{datasets}
