% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_amplification.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Amplification timing}
\usage{
x \%!in\% y
}
\arguments{
\item{cn_data}{Data frame containing copy number information.}

\item{multiplicity_data}{Data frame containing multiplicity information.}

\item{sample_id}{ID name or label for sample.}

\item{amplification_chrom}{Chromosome of the amplified region.}

\item{amplification_start}{Start position of the amplified region.}

\item{amplification_stop}{End position of the amplified region.}

\item{is_WGD}{logical: TRUE indicates that the sample has been whole genome duplicated.}
}
\value{
A data frame containing approximate timing of each amplification, and the most likely order of events.
}
\description{
Time individual amplification events for a locus that has been gained multiple times.
}
\examples{
sample1_time <- time_amplification(cn_data = sample1_cn, multiplicity_data = sample1_multiplicity, amplification_chrom = c("1"), amplification_start = c("100"), amplification_stop = c("100000"), is_WGD = FALSE);
}
