% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_amplification.R
\name{time_amplification}
\alias{time_amplification}
\title{Amplification timing}
\usage{
time_amplification(
  cn_data,
  multiplicity_data,
  sample_id,
  amplification_chrom,
  amplification_start,
  amplification_stop,
  is_WGD
)
}
\arguments{
\item{cn_data}{Data frame containing copy number information.}

\item{multiplicity_data}{Data frame containing multiplicity information.}

\item{sample_id}{ID name or label for sample.}

\item{amplification_chrom}{Chromosome of the amplified region.}

\item{amplification_start}{Start position of the amplified region.}

\item{amplification_stop}{End position of the amplified region.}

\item{is_WGD}{logical: TRUE indicates that the sample has been whole genome duplicated.}
}
\value{
A data frame containing approximate timing of each amplification, and the most likely order of events.
}
\description{
Time individual amplification events for a locus that has been gained multiple times.
}
