% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_amplification.R
\name{time_amplification}
\alias{time_amplification}
\title{Amplification timing}
\usage{
time_amplification(
  cn_data,
  multiplicity_data,
  mutation_data,
  muts_type = "All",
  sample_id,
  amplification_chrom,
  amplification_start,
  amplification_stop,
  is_WGD,
  genome
)
}
\arguments{
\item{cn_data}{Data frame containing copy number information derived from Battenberg.
The data frame must contain the following column names: "chr","startpos","endpos","nMaj1_A","nMin1_A"}

\item{multiplicity_data}{Data frame containing multiplicity information derived from DPClust.
The data frame must contain the following column names: "chr","end","no.chrs.bearing.mut"}

\item{mutation_data}{Data frame containing mutation data, including Ref and Alt alleles. Columns must be "chr","start","end","ref","alt". 
This file is required if muts_type = "All", as mutation context is required to 
identify clocklike mutations (C>T at CpG), and should contain all mutations. 
If muts_type == "SBS1 and SBS5" this file should contain only mutations that 
have been assigned to mutational signatures SBS1 or SBS5.}

\item{muts_type}{Must be one of "SBS1 and SBS5" or "All". Default is set to "All".  If set to "All", 
the mutation_data field must be supplied, and the code will be restricted to analysis of clock-like mutations (C>T mutations at CpG sites). If set to "SBS1 and SBS5", the multiplicity_data data frame should have been filtered so as to contain only mutations that have been attributed to mutational signatures SBS1 and SBS5.}

\item{sample_id}{ID name or label for sample. Character.}

\item{amplification_chrom}{Chromosome of the amplified region. Must not contain "chr" in the name.}

\item{amplification_start}{Start position of the amplified region. Numeric.}

\item{amplification_stop}{End position of the amplified region.  Numeric.}

\item{is_WGD}{logical: TRUE indicates that the sample has been whole genome duplicated.
FALSE indicates that the sample has not been whole duplicated.}

\item{genome}{Reference genome used. Must be one of "hg19" or "hg38".}
}
\value{
A data frame containing approximate timing of each amplification, and the most likely order of events.
}
\description{
Wrapper function that times individual amplification events for a locus that has been gained multiple times.
}
\examples{
demo_cn <- as.data.frame(matrix(nrow = 1, ncol = 5, data = c(1,50000,1000000,2,1)))
colnames(demo_cn) <- c("chr","startpos","endpos","nMaj1_A","nMin1_A")
demo_cn$startpos <- as.numeric(demo_cn$startpos)
demo_cn$endpos <- as.numeric(demo_cn$endpos)
demo_cn$nMaj1_A <- as.numeric(demo_cn$nMaj1_A)
demo_cn$nMin1_A <- as.numeric(demo_cn$nMin1_A)

demo_mult <- as.data.frame(matrix(nrow = 20, ncol = 3, data = c(rep("1",20),57682:57701,rep(2,5),rep(1,15))))
colnames(demo_mult) <- c("chr","end","no.chrs.bearing.mut")
demo_mult$end <- as.numeric(demo_mult$end)
demo_mult$no.chrs.bearing.mut <- as.numeric(demo_mult$no.chrs.bearing.mut)

demo_muts <- as.data.frame(matrix(nrow = 20, ncol = 5, data = c(rep("1",20),57682:57701,57682:57701,rep("C",15),rep("A",5),rep("T",15),rep("G",5))))
colnames(demo_muts) <- c("chr", "start", "end", "ref", "alt")
demo_muts$start <- as.numeric(demo_muts$start)
demo_muts$end <- as.numeric(demo_muts$end)

time_amplification(
  cn_data = demo_cn,
  multiplicity_data = demo_mult,
  mutation_data = demo_muts,
  muts_type = "All",
  sample_id = "test_sample",
  amplification_chrom = 1,
  amplification_start = 50100,
  amplification_stop = 1500000,
  is_WGD = TRUE,
  genome = "hg19"
)
}
